Fix spec clone method and usage of raw vs resolved spec (#2089)
This PR fixes an issue introduced in #2002, and the original issue #2002
was trying to address.

The original issue was that a cloned spec did not have properly resolved
references. #2002 fixed this incorrectly by cloning the resolved spec,
while the `Spec` initializer expects a raw spec.

This PR fixes this by cloning the raw spec, and passing the `base_uri`
required to resolve it along to the initializer of the new `Spec`
instance.

The swagger ui was also updated to use the resolved spec instead of the
raw spec.

Supersedes:
#1889
#2080

Fixes:
#1890
#1909
#2028 
#2029